+++ sample-apps/client/ts-quickstart/src/main.ts
@@ -1,28 +1,38 @@
 import './style.css';
 import { StreamVideoClient, User } from '@stream-io/video-client';
-import { renderParticipant } from './participant';
+import { decode } from 'js-base64';
+import { cleanupParticipant, renderParticipant } from './participant';
 import { renderControls } from './controls';
 import {
   renderAudioDeviceSelector,
+  renderAudioOutputSelector,
   renderVideoDeviceSelector,
+  renderVolumeControl,
 } from './device-selector';
 import { isMobile } from './mobile';
 
-const apiKey = import.meta.env.VITE_STREAM_API_KEY;
-const token = import.meta.env.VITE_STREAM_USER_TOKEN;
-const user: User = { id: import.meta.env.VITE_STREAM_USER_ID };
+const searchParams = new URLSearchParams(window.location.search);
+const extractPayloadFromToken = (token: string) => {
+  const [, payload] = token.split('.');
 
-const client = new StreamVideoClient({
-  apiKey,
-  token,
-  user,
-});
+  if (!payload) throw new Error('Malformed token, missing payload');
+
+  return (JSON.parse(decode(payload)) ?? {}) as Record<string, unknown>;
+};
+
+const apiKey = import.meta.env.VITE_STREAM_API_KEY;
+const token = searchParams.get('ut') ?? import.meta.env.VITE_STREAM_USER_TOKEN;
+const user: User = {
+  id: extractPayloadFromToken(token)['user_id'] as string,
+};
 
 const callId =
+  searchParams.get('call_id') ||
   import.meta.env.VITE_STREAM_CALL_ID ||
   (new Date().getTime() + Math.round(Math.random() * 100)).toString();
-const call = client.call('default', callId);
 
+const client = new StreamVideoClient({ apiKey, token, user });
+const call = client.call('default', callId);
 call.join({ create: true }).then(async () => {
   // render mic and camera controls
   const controls = renderControls(call);
@@ -38,6 +48,13 @@ call.join({ create: true }).then(async () => {
   } else {
     container.appendChild(renderVideoDeviceSelector(call));
   }
+
+  const audioOutputSelector = renderAudioOutputSelector(call);
+  if (audioOutputSelector) {
+    container.appendChild(audioOutputSelector);
+  }
+
+  container.appendChild(renderVolumeControl(call));
 });
 
 window.addEventListener('beforeunload', () => {
@@ -45,18 +62,23 @@ window.addEventListener('beforeunload', () => {
 });
 
 const parentContainer = document.getElementById('participants')!;
+call.setViewport(parentContainer);
 
 call.state.participants$.subscribe((participants) => {
-  const participantElements = participants.map((participant) =>
-    renderParticipant(call, participant),
-  );
+  // render / update existing participants
+  participants.forEach((participant) => {
+    renderParticipant(call, participant, parentContainer);
+  });
 
-  // Update UI
-  parentContainer.innerHTML = '';
-  participantElements
-    .flatMap((e) => [e.audioEl, e.videoEl])
-    .filter((el) => !!el)
+  // Remove stale elements for stale participants
+  parentContainer
+    .querySelectorAll<HTMLMediaElement>('video, audio')
     .forEach((el) => {
-      parentContainer.appendChild(el!);
+      const sessionId = el.dataset.sessionId!;
+      const participant = participants.find((p) => p.sessionId === sessionId);
+      if (!participant) {
+        cleanupParticipant(sessionId);
+        el.remove();
+      }
     });
 });